Add a unit test and fix some spelling
