+++ java/openmldb-batch/src/test/scala/com/_4paradigm/openmldb/batch/TestUnsafeRowWindowProject.scala
@@ -23,7 +23,7 @@ import org.apache.spark.sql.{Row, SparkSession}
 
 class TestUnsafeRowWindowProject extends SparkTestSuite {
 
-  test("TestUnsafeRowWindowProject") {
+  test("Test unsafeRow window project") {
 
     val spark = SparkSession
       .builder
