+++ java/openmldb-batch/src/test/scala/com/_4paradigm/openmldb/batch/end2end/TestWindowSkewOptAndParallelization.scala
@@ -22,10 +22,11 @@ import com._4paradigm.openmldb.batch.utils.SparkUtil
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.types.{DoubleType, IntegerType, StringType, StructField, StructType}
 
-class TestWindowSkewOptimizationWithUnion extends SparkTestSuite {
+class TestWindowSkewOptAndParallelization extends SparkTestSuite {
 
-  test("Test window union in INSTANCE_NOT_IN_WINDOW") {
+  test("Test end2end window skew optimization and window paralleization with union") {
 
+    getSparkSession.conf.set("spark.openmldb.window.parallelization", true)
     getSparkSession.conf.set("spark.openmldb.window.skew.opt", true)
     val spark = getSparkSession
     val sess = new OpenmldbSession(spark)
@@ -39,18 +40,18 @@ class TestWindowSkewOptimizationWithUnion extends SparkTestSuite {
         |     PARTITION BY t1.col1
         |     ORDER BY t1.col2
         |     ROWS_RANGE BETWEEN 3 PRECEDING AND CURRENT ROW
-        |     INSTANCE_NOT_IN_WINDOW)
+        |     )
         | limit 10;
         |""".stripMargin
 
     val data1 = Seq(
-      Row(5, 2, 1.1, "a"),
-      Row(5, 4, 2.2, "b"),
-      Row(5, 6, 3.3, "c"))
+      Row(5, 2, 2, "a"),
+      Row(5, 4, 4, "b"),
+      Row(5, 6, 6, "c"))
     val schema1 = StructType(List(
       StructField("col1", IntegerType),
       StructField("col2", IntegerType),
-      StructField("col3", DoubleType),
+      StructField("col3", IntegerType),
       StructField("col4", StringType)))
 
     val df1 = spark.createDataFrame(spark.sparkContext.makeRDD(data1), schema1)
@@ -58,13 +59,13 @@ class TestWindowSkewOptimizationWithUnion extends SparkTestSuite {
     sess.registerTable("t1", df1)
 
     val data2 = Seq(
-      Row(5, 1, 1.0, "e"),
-      Row(5, 3, 2.0, "f"),
-      Row(5, 5, 3.0, "g"))
+      Row(5, 1, 1, "e"),
+      Row(5, 3, 3, "f"),
+      Row(5, 5, 5, "g"))
     val schema2 = StructType(List(
       StructField("c1", IntegerType),
       StructField("c2", IntegerType),
-      StructField("c3", DoubleType),
+      StructField("c3", IntegerType),
       StructField("c4", StringType)))
 
     val df2 = spark.createDataFrame(spark.sparkContext.makeRDD(data2), schema2)
@@ -74,17 +75,55 @@ class TestWindowSkewOptimizationWithUnion extends SparkTestSuite {
     val outputDf = sess.sql(sqlText)
 
     val compareData = Seq(
-      Row(5, 2, 2.1),
-      Row(5, 4, 5.2),
-      Row(5, 6, 8.3))
+      Row(5, 2, 3),
+      Row(5, 4, 10),
+      Row(5, 6, 18))
 
     val compareSchema = StructType(List(
       StructField("col1", IntegerType),
       StructField("col2", IntegerType),
-      StructField("w1_col3_sum", DoubleType)))
+      StructField("w1_col3_sum", IntegerType)))
 
     val compareDf = spark.createDataFrame(spark.sparkContext.makeRDD(compareData), compareSchema)
 
     assert(SparkUtil.approximateDfEqual(outputDf.getSparkDf(), compareDf, false))
+  }
+
+  test("Test end2end window skew optimization") {
+
+    getSparkSession.conf.set("spark.openmldb.window.parallelization", true)
+    getSparkSession.conf.set("spark.openmldb.window.skew.opt", true)
+    val spark = getSparkSession
+    val sess = new OpenmldbSession(spark)
+
+    val data = Seq(
+      Row(1, "tom", 100, 1),
+      Row(2, "amy", 200, 2),
+      Row(3, "tom", 300, 3),
+      Row(4, "amy", 400, 4),
+      Row(5, "tom", 500, 5))
+    val schema = StructType(List(
+      StructField("id", IntegerType),
+      StructField("user", StringType),
+      StructField("trans_amount", IntegerType),
+      StructField("trans_time", IntegerType)))
+    val df = spark.createDataFrame(spark.sparkContext.makeRDD(data), schema)
+
+    sess.registerTable("t1", df)
+    df.createOrReplaceTempView("t1")
+
+    val sqlText ="""
+                   | SELECT sum(trans_amount) OVER w AS w_sum_amount FROM t1
+                   | WINDOW w AS (
+                   |    PARTITION BY user
+                   |    ORDER BY trans_time
+                   |    ROWS BETWEEN 10 PRECEDING AND CURRENT ROW);
+     """.stripMargin
+
+    val outputDf = sess.sql(sqlText)
+
+    val sparksqlOutputDf = sess.sparksql(sqlText)
+    // Notice that the sum column type is different for SparkSQL and OpenMLDB batch
+    assert(SparkUtil.approximateDfEqual(outputDf.getSparkDf(), sparksqlOutputDf, false))
   }
 }
